refactor(cli+textarea): textarea file split, handleAgentMessage extraction, overlay unify
Phase 2 Tasks 2.1+2.2 + Phase 3 Tasks 3.7+3.8:
- textarea.go split: textarea.go(core) + view.go + update.go + text.go + scroll.go (2128→5 files)
- cli_agent_msg.go: extracted handleCancelAck + tryFlushMessageQueue
- cli_update.go: extracted handleTokenRefresh/handleEasterEgg* handlers
- cli_view.go: renderActiveOverlay unifies 3 overlay if-blocks
